% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeardLengths_doc.R
\docType{data}
\name{BeardLengths}
\alias{BeardLengths}
\title{BeardLengths}
\format{A data frame containing 150 rows and 2 columns
\describe{
  \item{Ship}{(character) - The pirate's ship}
  \item{Beard}{(numeric) - The length of the pirate's beard in cm}
 }}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
BeardLengths
}
\description{
A dataframe containing the lengths of beards on 3 different pirate ships
}
\keyword{datasets}
