% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamonds_doc.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{diamonds}
\format{
A data frame containing 300 rows and 4 columns
\describe{
  \item{weight}{(numeric) - The weight of the diamond}
  \item{clarity}{(numeric) - The clarity of the diamond}
  \item{color}{(numeric) - The color shading of the diamond}
  \item{value}{The value of the diamond}
 }
}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
diamonds
}
\description{
A dataframe containing information about 150 diamonds sold at auction.
}
\keyword{datasets}
