% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examscores_doc.R
\docType{data}
\name{examscores}
\alias{examscores}
\title{examscores}
\format{A data frame containing 100 rows and 4 columns
\describe{
  \item{a}{(numeric) - Score on exam a}
  \item{b}{(numeric) - ...exam b}
  \item{c}{(numeric) - ...exam c}
  \item{d}{(numeric) - ...exam d}
 }}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
examscores
}
\description{
A dataframe containing the results of 4 exams given to 100 students. Each row represents a student, each column is a score on an exam
}
\keyword{datasets}

