% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_doc.R
\docType{data}
\name{movies}
\alias{movies}
\title{movies}
\format{
A data frame containing 5000 rows and 13 columns
\describe{
  \item{name}{Movie name}
  \item{rating}{MPAA rating}
  \item{genre}{Movie genre}
  \item{creative.type}{Creative type}
  \item{time}{Running time in minutes}
  \item{year}{Year of release}
  \item{production.method}{Production method}
  \item{sequel}{Was the movie a sequel? 1 = yes, 0 = no}
  \item{budget}{Movie budget (in $USD millions)}
  \item{revenue.all}{Gross worldwide revenue in $USD millions}
  \item{revenue.dom}{Domestic revenue in $USD millions}
  \item{revenue.int}{International revenue in $USD millions}
  \item{revenue.inf}{Inflation adjusted worldwide revenue in $USD millions}
 }
}
\source{
www.the-numbers.com
}
\usage{
movies
}
\description{
A dataframe containing information about the top 5000 grossing movies of all time.
}
\keyword{datasets}
