% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pirate.pal_function.R
\name{pirate.pal}
\alias{pirate.pal}
\title{pirate.pal}
\usage{
pirate.pal(palette = "random", action = "return", trans = 1)
}
\arguments{
\item{palette}{A string defining the color palette to use (see examples). To use a random palette, use "random"}

\item{action}{Either "return" to return a vector of colors, "s" to show the full palette, or "showall" to plot all palettes}

\item{trans}{A number in the interval [0, 1] indicating how transparent to make the colors}
}
\description{
This function provides a number of color palettes
}
\examples{
# Show all palettes

pirate.pal(palette = "all", action = "show")

# Show som palettes

pirate.pal(palette = "nemo", action = "show")
pirate.pal(palette = "espresso", action = "show")

# Scatterplot using the "up" palette

x <- rnorm(100)
y <- x + rnorm(100, 0, .3)

plot(x, y, pch = 16, col = sample(pirate.pal("up"), 100, replace = T))
}
\keyword{colors}

