% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/piratepalette_function.R
\name{piratepalette}
\alias{piratepalette}
\title{piratepalette}
\usage{
piratepalette(palette = "random", action = "return", trans = 0)
}
\arguments{
\item{palette}{A string defining the color palette to use (see examples). To use a random palette, use "random"}

\item{action}{Either "return" to return a vector of colors, "s" to show the full palette, or "showall" to plot all palettes}

\item{trans}{A number in the interval [0, 1] indicating how transparent to make the colors}
}
\description{
This function provides a number of color palettes
}
\examples{
# Show all palettes

piratepalette(palette = "all", action = "show")

# Show som2 palettes

piratepalette(palette = "nemo", action = "show")
piratepalette(palette = "espresso", action = "show")
piratepalette(palette = "bugs", action = "show")

# Plot some baloons using the "up" palette
x <- rnorm(100, 0)
y <- rnorm(100, 2, 1.3)
plot(1, xlim = c(-7, 7), ylim = c(-7, 7),
     xlab = "", ylab = "", type = "n", xaxt = "n", yaxt = "n", bty = "n")

rect(-2, -6, 2, -2)
polygon(c(-2, 0, 2),
        c(-2, 0, -2)
)
rect(-7, -7, -2, 100)
rect(2, -7, 7, 100)
rect(-.5, -6, .5, -4)
points(.3, -5)
}
\keyword{colors}

