% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirateplot_function.R
\name{pirateplot}
\alias{pirateplot}
\title{pirateplot}
\usage{
pirateplot(formula, data, line.fun = mean, pal = "appletv",
  back.col = gray(1), point.cex = 1, point.pch = 1, point.lwd = 1,
  cut.min = "", cut.max = "", width.min = 0.3, width.max = NA,
  bean.o = 0.5, point.o = 0.5, bar.o = 0.5, hdi.o = 0, line.o = 0.5,
  hdi.iter = 1000, jitter.val = 0.03, ylim = "", xlim = "", ...)
}
\arguments{
\item{formula}{(formula) A formula in the form dv ~ iv indicating the dv and iv to be plotted.}

\item{data, }{(dataframe) Data which to perform the beanplot on. This data can consist of dataframes, vectors and/or formulas. For each formula, a dataset can be specified with data=[dataset], and a subset can be specified with subset=[subset]. If subset/data arguments are passed, but there are not enough subset/data arguments, they are reused. Additionally, na.action, drop.unused.levels and xlev can be passed to model.frame in the same way. Also, parameters for axis and title can be passed.}

\item{line.fun}{(function) A function that determines how average lines and bar heights are determined (default is mean).}

\item{pal}{(string) A string indicating the color palette of the plot. Can be a single color, or the name of a palette in the piratepal() function (e.g.; "basel", "google", "southpark")}

\item{point.cex, }{point.pch, point.lwd (numeric) Numbers indicating the size, pch type, and line width of raw data points.}

\item{cut.min, }{cut.max (numeric) Optimal minimum and maximum values of the beans.}

\item{width.min, }{width.max (numeric) The minimum and maximum width of a bean.}

\item{bar.o, }{point.o, hdi.o, line.o, bean.o (numeric) A number between 0 and 1 indicating how opaque to make the bars, points, hdi band, average line, and beans respectively.}

\item{hdi.iter}{(integer) An integer indicating how many iterations to run when calculating the HDI.}

\item{jitter.val}{(numeric) A number indicaing how much to jitter the points. Defaults to 0.05.}

\item{...}{other arguments passed on to the plot function (e.g.; main, xlab, ylab, ylim)}

\item{backcol}{(string) A string indicating the color of the plotting background}
}
\description{
The pirateplot function creates an RDI plot (Raw data, Descriptive and Inferential statistic) pirate version of the fantastic beanplot function in the beanplot package. Just like a beanplot, pirateplot takes a discrete iv and a continuous dv, and creates a plot showing raw data, smoothed densities and central tendency. In addition, pirateplot adds the option for a 95% Highest Density Intervals (HDI), and has a few aesthetic differences preferred by pirates.
}
\examples{



# Pirateplots of the ChickWeight dataframe

par(mfrow = c(2, 3))

# Plot 1: Colorful pirateplot

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "All Elements",
          hdi.o = .7,
          bar.o = .1,
          line.o = 1
)


# Plot 2: Black and white, turn down the beans and points,
#   Turn up the HDI

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "B&W with 95\% HDI",
          pal = "black",
          bar.o = .1,
          hdi.o = .5,
          point.o = .1,
          bean.o = 0
)

# Plot 3: Turn everything off except the bars

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "Super boring bars",
          pal = "black",
          bar.o = .9,
          hdi.o = 0,
          point.o = 0,
          bean.o = 0,
          back.col = "white",
          line.o = 0
)


# Plot 3: Only points, but larger with more jitter

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "Large, jittered points\\nGoogle colors",
          pal = "google",
          bar.o = 0,
          hdi.o = 0,
          point.o = .1,
          bean.o = 0,
          back.col = "white",
          line.o = 1,
          point.pch = 16,
          point.cex = 1.5,
          jitter.val = .1
)

# Plot 3: Beans and HDI

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "Beans and HDI\\nSouth Park colors",
          pal = "southpark",
          bar.o = 0,
          hdi.o = .5,
          point.o = 0,
          bean.o = 1,
          back.col = "white",
          line.o = 0,
          point.pch = 16
)


# Plot 3: Beans and HDI

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "HDI and points\\nDark background",
          pal = "ipod",
          bar.o = .2,
          point.o = .3,
          bean.o = 0,
          back.col = gray(.3),
          line.o = 0,
          hdi.o = .8,
          point.pch = 16
)

par(mfrow = c(1, 1))







}
\keyword{plot}

