% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pirates_doc.R
\docType{data}
\name{pirates}
\alias{pirates}
\title{pirates}
\format{A data frame containing 1,000 rows and 11 columns
\describe{
  \item{id}{price, in US dollars}
  \item{sex}{weight of the diamond, in carats}
  \item{headband}{a yes or no string value indicating whether or not a pirate wears a headband}
  \item{age}{age of the pirate in years}
  \item{college}{the college the pirate went to}
  ...
}}
\source{
1645 annual international pirate meeting
}
\usage{
pirates
}
\description{
A dataset containing the results of a survey and combat skills of 1,000 pirates
diamonds.
}
\keyword{datasets}

