% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pircharter_doc.R
\docType{data}
\name{pircharter}
\alias{pircharter}
\title{pircharter}
\format{A data frame containing 1000 rows and 10 columns
\describe{
  \item{company}{(string) - The charter company: JoRo = Jolly Roger, BmcB = Boaty McBoat, MiPa = Millenium Parrot}
  \item{destination}{(string) - The destination of the charter}
  \item{time}{(numeric) - The travel time of the ship in hours}
 }}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
pircharter
}
\description{
A dataframe containing travel times of chartered ships from three pirate companies to three different destinations.
}
\examples{




}
\keyword{datasets}

