% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poopdeck_doc.R
\docType{data}
\name{poopdeck}
\alias{poopdeck}
\title{poopdeck}
\format{
A data frame containing 300 rows and 4 columns
\describe{
  \item{day}{(factor) - The day that the poop deck was cleaned (1 through 10000)}
  \item{cleaner}{(string) - The cleaning solution used}
  \item{type}{(string) - The type of poop being cleaned}
  \item{time}{(numeric) - The amount of time (in minutes) the cleaning took.}
 }
}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
poopdeck
}
\description{
A dataframe containing the amount of time it took to clean both pirate and shark poop from the poop deck using one of three different cleaning solutions
}
\keyword{datasets}
