% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodev_function.R
\name{recode.v}
\alias{recode.v}
\title{recodev}
\usage{
recode.v(original.vector, old.values, new.values, others = NULL)
}
\arguments{
\item{original.vector}{A vector you want to recode}

\item{old.values}{A vector of length M.}

\item{new.values}{A vector of length M.}

\item{others}{An optional value indicating what to convert all values in original.vector that are not found in old.values.}
}
\description{
This function takes a vector original.vector, and converts all values in a vector old.values to the values in a new vector new.values.
}
\examples{

x <- c("y", "y", "XSF", "y", "0", "X", "0", "0", "y", "n", "0", "1", "1")
recode.v(original.vector = x,
         old.values = c("y", "1", "n", "0"),
         new.values = c(1, 1, 0, 0)
)


x <- c("y", "y", "XSF", "y", "0", "X", "0", "0", "y", "n", "0", "1", "1")
recode.v(original.vector = x,
         old.values = c("y", "1", "n", "0"),
         new.values = c(1, 1, 0, 0),
         others = NA
)


}

