% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tanny_function.R
\name{tanny_function}
\alias{tanny_function}
\title{Tanny function}
\usage{
tanny_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
tanny_function()
}
\keyword{cats}

