% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/transparent_function.R
\name{transparent}
\alias{transparent}
\title{transparent function}
\usage{
transparent(orig.col = "red", trans.val = 1, maxColorValue = 255)
}
\arguments{
\item{orig.col}{The original color to be made transparent. Can be specified as a string or a vector of rgb values}

\item{trans.val}{A number in the interval [0, 1] indicating how transparent to make the color.}

\item{maxColorValue}{The maximum color value (only used when orig.col is an rgb vector)}
}
\description{
This function takes a standard color as an argument and returns a
transparent version of that color
}
\examples{
transparent()
}
\keyword{colors}

