% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auction_doc.R
\docType{data}
\name{auction}
\alias{auction}
\title{auction}
\format{A data frame containing 1000 rows and 8 columns
\describe{
  \item{cannons}{(integer) The number of cannons on the ship}
  \item{rooms}{(integer) The number of rooms on the ship}
  \item{age}{(numeric) The age of the ship in years}
  \item{style}{(string) The style of the ship - either modern or classic}
  \item{condition}{(integer) The condition of the ship on a scale of 1 to 10}
  \item{weight}{(numeric) The weight of the ship}
  \item{color}{(string) The color of the ship}
  \item{price}{(numeric) The selling price of the ship (in gold pieces, obviously)}
  \item{sold}{(binary) The pre-sale predicted value of the ship according to Jack's Blue Book (JBB)}
}}
\source{
2015 annual pirate auction in Portland Oregon
}
\usage{
auction
}
\description{
A dataframe containing data from 1000 ships sold at a pirate auction.
}
\examples{
pirateplot(dv.name = "price",
iv.name = "condition",
data = auction
)
}
\keyword{datasets}

