% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BeardLengths_doc.R
\docType{data}
\name{BeardLengths}
\alias{BeardLengths}
\title{BeardLengths}
\format{A data frame containing 150 rows and 2 columns
\describe{
  \item{Ship}{A string indicating which ship the pirate is stationed on}
  \item{Beard}{The length of the pirate's beard in cm}
}}
\source{
2015 annual international pirate meeting at the Bodensee in Konstanz, Germany
}
\usage{
BeardLengths
}
\description{
A dataframe containing the ship and beard lengths of 150 pirates.
}
\keyword{datasets}

