% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/hsurvey_doc.R
\docType{data}
\name{hsurvey}
\alias{hsurvey}
\title{Survey of pirates}
\format{A data frame containing results from a survey of
\describe{
  \item{sex}{price, in US dollars}
  \item{age}{weight of the diamond, in carats}
  ...
}}
\source{
1645 annual international pirate meeting
}
\usage{
hsurvey
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

