% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_doc.R
\docType{data}
\name{movies}
\alias{movies}
\title{movies}
\format{A data frame containing 5,000 rows and 14 columns
\describe{
  \item{name}{movie name}
  \item{boxoffice.total}{Total box office revenue (in USD)}
  \item{boxoffice.domestic}{Domestic box office revenue (in USD)}
  \item{boxoffice.international}{International box office revenue (in USD)}
  \item{dvd.domestic}{DVD revenue (in USD)}
  \item{budget}{Production budget (in USD)}
  \item{rating}{MPAA rating}
  \item{genre}{Movie genre}
  \item{creative.type}{Type of movie}
  \item{time}{Running time (in minutes)}
  \item{year}{Release year}
  \item{production.method}{How the movie was produced}
  \item{sequel}{Was the movie a sequel?}
  \item{boxoffice.domestic.inflationadj}{Inflation adjusted domestic box office revenue (in USD)}

  
  ...
}}
\source{
1645 annual international pirate meeting
}
\usage{
movies
}
\description{
A dataset containing the 5,000 top grossing movies (of all time?)
}
\keyword{datasets}

