% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pirate.colors_function.R
\name{pirate.colors}
\alias{pirate.colors}
\title{transparent function}
\usage{
pirate.colors(palette = "random", action = "return", trans = 1)
}
\arguments{
\item{palette}{A string defining the color palette to use (see examples). To use a random palette, use "random"}

\item{action}{Either "return" to return a vector of colors, "plot" to plot the full palette, or "showall" to plot all palettes}
}
\description{
This function takes a standard color as an argument and returns a
transparent version of that color
}
\examples{
# Show all palettes

palette.names <- c("ipod", "nyt", "espresso", "info", "info2", "stork", "google", "drugs",
"goldfish", "provoking", "emo", "cake", "pancake", "lubitel")

plot(1, xlim = c(0, 10),
ylim = c(0, 1))
}
\keyword{colors}

