% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirateplot_function.R
\name{pirateplot}
\alias{pirateplot}
\title{pirateplot}
\usage{
pirateplot(formula, data, line.fun = mean, pal = "appletv",
  back.col = gray(1), point.cex = 1, point.pch = 1, point.lwd = 1,
  cut.min = "", cut.max = "", width.min = 0.3, width.max = NA,
  bean.o = 0.5, point.o = 0.5, bar.o = 0.5, hdi.o = 0, line.o = 0.5,
  hdi.iter = 1000, jitter.val = 0.03, ylim = "", xlim = "", ...)
}
\arguments{
\item{formula}{(formula) A formula in the form dv ~ iv indicating the dv and iv to be plotted.}

\item{data, }{(dataframe) Data which to perform the beanplot on. This data can consist of dataframes, vectors and/or formulas. For each formula, a dataset can be specified with data=[dataset], and a subset can be specified with subset=[subset]. If subset/data arguments are passed, but there are not enough subset/data arguments, they are reused. Additionally, na.action, drop.unused.levels and xlev can be passed to model.frame in the same way. Also, parameters for axis and title can be passed.}

\item{line.fun}{(function) A function that determines how average lines and bar heights are determined (default is mean).}

\item{pal}{(string) A string indicating the color palette of the plot. Can be a single color, or the name of a palette in the piratepal() function (e.g.; "basel", "google", "southpark")}

\item{point.cex, }{point.pch, point.lwd (numeric) Numbers indicating the size, pch type, and line width of raw data points.}

\item{cut.min, }{cut.max (numeric) Optimal minimum and maximum values of the beans.}

\item{width.min, }{width.max (numeric) The minimum and maximum width of a bean.}

\item{bar.o, }{point.o, hdi.o, line.o, bean.o (numeric) A number between 0 and 1 indicating how opaque to make the bars, points, hdi band, average line, and beans respectively.}

\item{hdi.iter}{(integer) An integer indicating how many iterations to run when calculating the HDI.}

\item{jitter.val}{(numeric) A number indicaing how much to jitter the points. Defaults to 0.05.}

\item{...}{other arguments passed on to the plot function (e.g.; main, xlab, ylab, ylim)}

\item{backcol}{(string) A string indicating the color of the plotting background}
}
\description{
The pirateplot function creates an RDI plot (Raw data, Descriptive and Inferential statistic) pirate version of the fantastic beanplot function in the beanplot package. Just like a beanplot, pirateplot takes a discrete iv and a continuous dv, and creates a plot showing raw data, smoothed densities and central tendency. In addition, pirateplot adds the option for a 95% Highest Density Intervals (HDI), and has a few aesthetic differences preferred by pirates.
}
\examples{



# Plot 1: Terrible Barplot

mean.agg <- aggregate(formula = weight ~ Diet,
                     data = ChickWeight,
                     FUN = mean
)

barplot(height = mean.agg$weight,
       names.arg = mean.agg$Diet,
       xlab = "Diet", ylab = "Weight (mean)",
       main = "Barplot of ChickWeight"
)

# Plot 2: Better Boxplot

boxplot(formula = weight ~ Diet,
       data = ChickWeight, xlab = "Diet",
       ylab = "Weight", main = "Boxplot of ChickWeight"
)


# Plot 3: Colorful pirateplot

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          main = "Color Pirateplot")

# Plot 4: Black and white pirateplot with smaller points
#   and different transparencies

pirateplot(formula = weight ~ Diet,
          data = ChickWeight,
          pal = "black",
          bar.o = 1,
          bean.o = .1,
          point.cex = .7,
          main = "B&W Pirateplot"
)

# Plot 5: Pirateplot with a 95\% HDI using
#   the South Park palette

pirateplot(formula = weight ~ Diet,
          pal = "southpark",
          main = "Pirateplot of Chicken Weights\\nwith 95\% HDI and South Park palette",
          data = ChickWeight,
          point.cex = .7,
          add.hdi = T
)







 # Some made-up survey data

PhoneData <- data.frame("Phone" = rep(c("iPhone", "Android", "Windows"), each = 50),
                       "Age" = c(rnorm(50, mean = 25, sd = 4),
                                 rnorm(50, mean = 20, sd = 3),
                                 rnorm(50, mean = 30, sd = 3)
                                 ),
                       stringsAsFactors = F
                       )

 pirateplot(formula = Age ~ Phone,
         data = PhoneData,
         width.min = .45,
         pal = "black",
         main = "Age of Phone Users"
         )







}
\keyword{plot}

